requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
httpx>=0.28.0
snowflake-connector-python>=3.12.0
pydantic>=2.10.0
pydantic-settings>=2.7.0
python-dotenv>=1.0.0

# Testing
pytest>=8.3.0
pytest-asyncio>=0.25.0
respx>=0.22.0
